requirements.txt
playwright==1.19.0
pytest-playwright==0.2.3
requests==2.26.0
